Moved `Closest` out into the `types` module
